requirement.txt
# Core Telegram library
pyrogram

# Untuk performa asyncio yang lebih cepat
uvloop

# Database driver untuk MongoDB
motor

# Untuk membaca variabel dari file .env
python-dotenv